Increase line length limit to 120
The 80 character line limit is a standard which has it's roots in a
time past. Specifically this rule came about from the time of punch cards
and then small screens. That's not to say we shouldn't have a line length
limit. You could put all of a program in one line, that wouldn't be fun
to read / maintain. The length of 120 is gaining traction, but it's kinda
a "magic number" in that I'm not sure we have logical proof of why 120
should be the number, but it seems reasonable.