Make `interfaces_version` a direct attribute
`interfaces_version` has to be a direct attribute of the `SemanticManifest`
for deserialization to work properly. With it being a direct attribute, we
had to begin using the pydantic validator to ensure the version gets set
on `SemanticManifest` instantiation.
Note: Callum McCann authored this commit, but I (Quigley Malcolm) have updated
the commit message.