Make type_enums importable from the type_enum module root
In core I've become slightly annoyed (with myself) by having to designate the
separate import paths for each type_enum when I'm generally importing
multiple in a given file. This change makes it so that one only needs to
specify the path as far as `dbt_semantic_interfaces.type_enums` to import
any given type_enum.