Make protocols importable from the protocols module root
In core I've become slightly annoyed (with myself) by having to designate the
separate import paths for each protocol when I'm generally importing
multiple in a given file. This change makes it so that one only needs to
specify the path as far as `dbt_semantic_interfaces.protocols` to import
any given protocol.

Note: This is non breaking as the old import paths still work.