Drop `create_metric` from the `Measure` protocol
The attribute `create_metric` is a parse time convenience attribute.
It's not something that MetricFlow will take action on. That is to say
for those implementing parsing, they may want to have something akin
to `create_metric` for the ergonomics of their parsing schema, but is
not something actionable when passing the objects to other systems
because the action that should be caused should have already resulted
in the creation of a metric before a manifest gets passed.