Update `checked_agg_time_dimension_for_measure` to only require a `MeasureReference`
Previously we were taking in a fully realized `Measure` object as a parameter
to `checked_agg_time_dimension_for_measure`.  By switching to taking in only
a `MeasureReference` the function now guarantees that the referenced `Measure`
belongs to the `SemanticModel`. This has the unfortunate side affect of preforming
an additional list based lookup, but if this ever becomes an issue, we'll know
what we can do.