Support for `Dimension.grain(...)` in where/filter (#152)
The object of this PR is to allow for syntax like this in the where parameter & filter field {{ Dimension('metric_time').grain('WEEK') }} > 2023-09-18. To accomplish this, I refactored WhereFilterParser to use the QueryInterface* classes such as QueryInterfaceDimension. This has been done in MF already for WhereSpecFactory.

Now that MF, MFS, and DSI all implement the same QueryInterface* classes, there will be a lessened likelihood of drift between these codebases. I plan on removing the QueryInterface* classes from MF and having that codebase reference the protocols in the codebase instead.