Add test cases for PydanticWhereFilter parsing
The WhereFilter implementation has a more complicated parser than
most other Pydantic implementation objects, because where filters
can be expressed as bare strings in the YAML, and for various
historical reasons we have had to allow the parser to process either
raw string inputs, properly serialized objects, or actual python
instances.

With the impending switch to WhereFilterIntersection objects, which
will need to parse all of this and more, some tests documenting and
enforcing the expected behavior for these classes seemed in order.