Add WhereFilterIntersection model object
We are converting the filter: parameters accepting WhereFilter
protocols in the semantic manifest to instead accept a set of
WhereFilters which will be applied as an intersection.

The cleanest way to do this while retaining backwards compatibility
is to add a new protocol for a WhereFilterIntersection and a corresponding
Pydantic model object implementation that consolidates all of the parsing
logic updates necessary to handle legacy semantic manifests.

This also has the advantage of allowing for easy extensions to these
filters for any other properties we might wish to apply to the full
set (e.g., predicate pushdown configuration options).

This commit simply adds the protocol, a working implementation, and
the additional test cases to demonstrate theoretical backwards compatibility.
No direct YAML handling has been tested yet, as the YAML parser has not been
hooked up to the new protocol layout.