Update SavedQuery objects to WhereFilterIntersection
For the sake of consistency, and everybody's sanity, we should
have the new things use the same constructs as the old things.

While the original Sequence[WhereFilter] type is nominally fine, it
would become rather annoying to have to deal with lists in some places
and objects containing lists in others, especially when collecting
call parameter sets.