Abstract duplicate logic within `_validate_top_level_objects` to helper function
We were doing a bunch of duplicate logic in `_validate_top_level_objects` that
could be abstracted out. Unfortunately abstracting it out required generating
the contexts for validatin issues in advance. It's not the prettiest, but it
was necessary. Ideally the context building should happen in the helper method
itself, however this isn't currently possible. We'd either need to have a
generic context, be able to identify which context needs to be built, or have
a generalized context class.