Fix group/ignore rules in dependabot.yml
The original "groups" approach caused inscrutable errors. In digging further I discovered that it also did not do what I thought it did - by "excluding" those dependencies from the group I was opening them up to individual PR updates.

This removes the grouping since I can't be bothered to figure out how to make it work, and adds the proper `ignore` key for the dependencies we do not wish to auto-update. The latter can also be updated via @ commands to dependabot but I prefer to keep them catalogued here.