Add a pydantic shim for handling pydantic 1 vs pydantic 2
We are going to need to support pydantic 1 for some time. We think currently
that we'll have to support it for atleast another year. However, we also
need to begin supporting pydantic 2. Thus we need this shim to have the
flexibility of supporting both.

I have also run `hatch build` and ensured that this shim _does_ get included
in the resulting tarball. In the upcoming commits, we'll switch over direct
pydantic imports to instead use this shim, and then loosen the version in the
pyproject config.