Rename `pydantic_shim.py` to `dsi_pydantic_shim` to reduce collision chance
The pydantic shim we provide alongside the dbt-semantic-interfaces package is
installed outside of the dbt-semantif-interfaces by pip. That is when the
wheel is installed the venv lib looks something like

```
venv/
  ...
  lib/
    ...other packages
    dbt-semantic-interfaces
    ...other packages
    pydantic_shim.py
    ...
```

The problem with this is that if another package is installed which also provides
a pydantic shim, we risk a name collision. Put another way if `package_foo` and
`dbt-semantic-interfaces` both provide a `pydantic_shim.py` alongside their package,
then pip will likely be unhappy. To avoid this we've renamed our shim to
`dsi_pydantic_shim.py` to denote our shim as being specific to dbt-semantic-interfaces.