Update pyproject.toml to include pydantic shim in wheel distribution building
Previously we weren't specifying what was to be included in the wheel build,
thus the wheel distribution was going through the default paths, which were
as documented [here](https://hatch.pypa.io/1.9/plugins/builder/wheel/#default-file-selection)

* <NAME>/__init__.py
* src/<NAME>/__init__.py
* <NAME>.py
* <NAMESPACE>/<NAME>/__init__.py

This meant that our pydantic shim was being excluded. This commit alters
the pyproject.toml to now begin specifying the packages to be included
in the wheel distribution, and notably adds our pydantic shim to it.