Allow customization of python version in env helper
We have a composite GitHub action for installing Python and
Hatch, but it is pegged to Python 3.9 whereas we'd typically want
to either test our minimum version (currently 3.8) or a matrix of
supported versions.

This parameterizes the action so we can pass in the version we
would like to test, and updates existing callers accordingly.