Update actions to use a version matrix with the composite setup
The composite action for setting up a python env with an interpreter
and hatch install now allows parameterized configuration,
meaning we can use a python version matrix across all of our actions.

This commit moves the pytest action to use the env setup helper,
and updates the other actions to run a test matrix against our
endpoint versions (currently 3.8 and 3.12) under the assumption
that things working on the endpoints will also work in the middle.
While we could test all versions, in practice the actions with this
smaller test surface wouldn't really benefit from this, and any
developer running into problems there should be able to fix and test
version-specific problems independently.