Implement new time spine YAML keys
As we work to support sub-daily granularities, users will be able to set up multiple time spines at different granularities. Currently, these time spines are not configured in YAML, and these fields are set in core instead and show up in the manifest. That means we should be able to change the names here as long as we also update them in core and anywhere they are handled downstream, such as MF or MFS.