Add `default_granularity` to metric spec (#291)
Resolves #290

### Description
Now that we are enabling sub-daily granularities, the current behavior
of `metric_time` will likely be unexpected to users. If they change any
of their time dimensions to use a grain smaller than DAY, `metric_time`
will default to that grain. To avoid this confusion, add a parameter
that allows users to choose which grain they want their metric to
default to. This will default to DAY if not set to avoid the unexpected
behavior change. (If DAY is smaller than the defined granularity, will
default to the smallest available granularity.)


### Checklist

- [x] I have read [the contributing
guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md)
and understand what's expected of me
- [x] I have signed the
[CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [x] I have run `changie new` to [create a changelog
entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)