Rename `Metric.default_granularity` field to `Metric.time_granularity` (#310)
### Description
After some discussion with product and dev experience, we've decided to
change the name of this field. It hasn't shipped to core yet so is not a
breaking change. The name change is intended to remove the word
"default", which should be implied for a YAML field. This is more
consistent with other dbt YAML objects, which have a default in YAML
that can be overridden at query time. The name `time_granularity` is
also more consistent with the field we have on time dimensions.
Everything in this PR is just a find & replace operation.

### Checklist

- [x] I have read [the contributing
guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md)
and understand what's expected of me
- [x] I have signed the
[CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [x] I have run `changie new` to [create a changelog
entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)