Streamline object-builder text parsing (#322)
### Description

When parsing the Jinja template in where-filters, there is currently a
set of factory and factory-specific classes to implement a specific
behavior. I've noticed that there are multiple implementations of those
classes (if you add the ones in this repo and dependent repos), and it
seemed like there was an opportunity to streamline the structure to
handle the update for saved-queries and other changes down the road.

This PR adds a `QueryItemDescription` that encapsulates what the user
has specified in the text input, and also adds a
`QueryItemTextProcessor` that allows you to:

* Find all the `QueryItemDescription`s in a template.
* Render a template for a given `QueryItemDescription` using a
configurable format.
* Configure what builder methods (e.g. `.descending()`) are allowed in
which cases.

These features seem like it would support existing use cases, but
tagging @DevonFulcher to confirm.

Also, please view by commit. The first commit includes the interface, so
if there are issues there, it would be better to get those resolved
before updating the later commits.

### Checklist

- [x] I have read [the contributing
guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md)
and understand what's expected of me
- [x] I have signed the
[CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [x] I have run `changie new` to [create a changelog
entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)