Handle default of None properly for cumulative_type_params
dbt-core parses YAML differently than DSI. There, the YAML dict for a metric ends up with  if that key is empty. Then when we call type_params.get(cumulative_type_params, {}), the result is None because the key does exist in the dict. This results in a key error when we finally call .get(grain_to_date). Thus, we must set the default value outside of .get().