Saved query order by support (#382)
### Description
Currently, the `order_by` param in saved query YAML doesn't work in
MetricFlow. This PR is part of the work to unblock that.
A few things to get there:
- Supports differentiating between whether you're parsing Jinja in an
order by or elsewhere. This is necessary because the `descending` method
should only be supported in the order by, and `group_by` should not be
required for order by metrics.
- Renames a few classes & one file to be more generic - to reflect the
fact that the Jinja may not only be in the where filter
- Adds the descending attribute to the call parameter set objects, which
is needed to pass that attribute around in MF

I highly recommend reviewing by commit!

### Checklist

- [x] I have read [the contributing
guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md)
and understand what's expected of me
- [x] I have signed the
[CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [x] I have run `changie new` to [create a changelog
entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)