Add measures fields to metrics classes (#390)
Towards Issue
[#387](https://github.com/dbt-labs/dbt-semantic-interfaces/issues/387)
Internal SL-4116

### Description

This PR stack is part of a larger YAML spec re-write that simplifies a
number of frequently confusing parts of the semantic layer. The biggest
change for this effort is making Simple Metrics (i.e. metrics with
type=SIMPLE) into the starting point for your semantic layer rather than
Measures, as the line between these two is often blurry and confusing
for users and seems a bit arbitrary, and that's the area that this group
of PRs is meant to start to address.

The first step is to add new functionalities to various classes and make
sure project stakeholders all have the same agreed-upon understanding of
how the internals will look.

For the first few PRs, just adding the new fields:
* (this PR) #390 adds new fields pulled from measures to Simple Metrics
to allow them to function without a separately-defined measure
* #391 - the new spec requires that certain fields move from the
consuming metric to the input metric, and this adds them there.
* #392 - some metrics consume measures today. This PR adds the ability
to consume from metrics to the those. (We'll add validation limiting
which metrics can be consumed later.)


NOTE: These changes must NOT be released before we finish updating the
validations (so folks don't create broken/unsupported configurations and
hit strange failures) and also update our other offerings to be able to
handle the updated spec!

### More details about this PR:
For schema notes, dbt contributors can reference:
* Dave's scratch schema
* Dave's yaml examples

In the YAML, a metric might look like:
```
metrics: # simple metrics only (could also include advanced metrics that are only derived from metrics on the same single model)

      - name: my_simple_metric # required
        description: cool metric # optional
        label: my label # optional
        type: simple  # required -- simple | cumulative | ratio | derived | conversion 
        agg: sum # required sum | max | min | average | median | count_distinct | percentile, and sum_boolean (use existing enum from DSI)
        expr: case when is_a then 1 else 0 end # optional for simple, defaults to name of metric
```
Note that the metric will now have new fields! But while the new look is
significantly flattened, we'll move the new measure-like fields into a
separate bag of params object when we parse it so that we can keep
things organized in our code.

### Checklist

- [x] I have read [the contributing
guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md)
and understand what's expected of me
- [x] I have signed the
[CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [x] I have run `changie new` to [create a changelog
entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)