Add input-altering fields directly to simple metrics (#391)
Towards Issue
[#387](https://github.com/dbt-labs/dbt-semantic-interfaces/issues/387)
Internal SL-4116

### Description

This PR stack is part of a larger YAML spec re-write that simplifies a
number of frequently confusing parts of the semantic layer. The biggest
change for this effort is making Simple Metrics (i.e. metrics with
type=SIMPLE) into the starting point for your semantic layer rather than
Measures, as the line between these two is often blurry and confusing
for users and seems a bit arbitrary, and that's the area that this group
of PRs is meant to start to address.

The first step is to add new functionalities to various classes and make
sure project stakeholders all have the same agreed-upon understanding of
how the internals will look.

For the first few PRs, just adding the new fields:
* #390 adds new fields pulled from measures to Simple Metrics to allow
them to function without a separately-defined measure
* (this PR) #391 - the new spec requires that certain fields move from
the consuming metric to the input metric, and this moves them there.
* #392 - some metrics consume measures today. This PR adds the ability
to consume from metrics to the those. (We'll add validation limiting
which metrics can be consumed later.)

NOTE: These changes must NOT be released before we finish updating the
validations (so folks don't create broken/unsupported configurations and
hit strange failures) and also update our other offerings to be able to
handle the updated spec!

### More details about this PR:
For schema notes, dbt contributors can reference:
* Dave's scratch schema
* Dave's yaml examples

In the YAML, a metric might look like:
```
      - name: my_simple_metric_with_configs_and_non_additive_dimension
        description: cool cumulative metric # optional
        label: my label # optional
        config:
          join_to_timespine: true # bool, if true, will join to timespine table
          fill_nulls_with: 0 # any, just what it sounds like! 
          meta: # optional, struct
            key: value 
          # optional advanced settings for any metric type
        non_additive_dimension: 
          name: my_derived_dimension # str, must be a dimension 
          window_choice: max # enum only min or max (might be in dbt-semantic-interfaces)
          window_groupings: 
            - my_entity_column # list of entity names
       # note: we include new fields for cumulative metrics in #392
      - name: my_cumulative_metric
        description: cool cumulative metric # optional
        label: my label # optional
        type: cumulative  # required -- simple | cumulative | ratio | derived | conversion 
        window: 1 week (interval) # optional 
        grain_to_date: hour # optional, enum of TimeGranularity
        period_agg: first # optional, enum of first | last | average
        metric: my_simple_metric # required, must refer to metric name or metric dict
```
Note that the simple metric taking the place of a measure now now holds
settings like `fill_nulls_with` and `join_to_timespine` rather than the
consuming metric (a cumulative metric in this case).

### Checklist

- [x] I have read [the contributing
guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md)
and understand what's expected of me
- [x] I have signed the
[CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [x] I have run `changie new` to [create a changelog
entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)