Add model specifier to metrics (#396)
Towards Issue
[#387](https://github.com/dbt-labs/dbt-semantic-interfaces/issues/387)
Internal SL-4116

### Description

This PR does two things. The MAIN thing it does is add a "model" field
to the new `PydanticMetricAggregationParams` param for simple metrics
that do not require a measure.

I *think* this was not initially included in the spec updates because
the model is implicitly included for measures, but since metrics can be
defined at the manifest level, there is no implicit definition for the
model that is being used, so we don't actually know what we're even
building a metric from.

The second change is moving the `join_to_timespine` and
`fill_nulls_with` from `PydanticMetricAggregationParams` to the metric
itself. `PydanticMetricAggregationParams` refers to a whole group of
related fields only used when we are using a simple metric without a
measure and which CANNOT be used without replacing a measure. However,
these fields CAN be used on any simple metric, so it's better if we
don't group them together. (This came up because I didn't want to make
"model" optional here, and I would have to if we left those two fields
in the `PydanticMetricAggregationParams` and then add more validations
and etc etc yuck.)

### Checklist

- [x] I have read [the contributing
guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md)
and understand what's expected of me
- [x] I have signed the
[CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for
this PR
- [x] I have run `changie new` to [create a changelog
entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)