Change "cumulative input is required" error to be a warning (#412)
Towards #387

### Description

Relaxes the validation error for providing no input to a cumulative metric (no measure nor metric) to be a warning instead of an error to avoid breaking changes.

Context: previously, we weren't checking this, so making this an error could break existing manifests that work in general and are being used to run jobs other than this cumulative metric.  (The metric will not work in metricflow without an input.)

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [x] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)