Relax errors for having both metric and measure inputs to be warnings (#419)
Towards #387

### Description

In order to be backwards-compatible, the transformations that will add metrics in place of measures can't actually remove the measure inputs.  That would mean that we'd be producing a manifest that would not pass validation if re-validated, which is a really nasty state to be in.  To address this, we relax the errors I added recently to be warnings in this specific case, and we add a little clarity about what will happen as a result of these warnings (i.e. the measure input will just be ignored in MF).

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [x] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)