Transform measures with create_metric true to actual metrics (#417)
Towards #387
Internal SL-4209

### Description

Since we're eventually transforming away measures entirely and replacing them with SImple metrics with a superset of their parameters and capabilities, this is a logical place to start.

Previously, setting `create_metric: True` in your measure yaml definition would cause us to create a simple metric with that measure as an input for you.

Now, we'll create a simple metric with all the correct settings to replace that measure.  This should not impact the final queries executed in MetricFlow.

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [x] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)