Add test for existing boolean aggregation transformation (#428)
Towards #387

### Description

Downstream (i.e. up-stack), we need to refactor the BooleanMeasureAggregationRule transformation logic a bit, but we don't have any tests in place to make sure this is working correctly.

This test canonizes the current behavior of the transformation so that we know the behavior is unchanged in subsequent PRs.

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [ ] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)