Add convert-count rule for metrics, too (#431)
Towards #387

### Description

We already have a transformation to turn COUNT measures into SUMs with useable `expr` values.  Since we're replacing measures with simple metrics, this PR adds a transformation to do the same thing for simple metrics with COUNT agg values.

We also do a little renaming and clean up for the existing (as of #430 , which is also in review) tests for the existing measure transformation.

Note: I didn't move the file for this one as I did for the boolean measure transformations because the file name was already measure-agnostic.

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [x] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)