Add tests for convert median transformation for measures (#432)
Towards #387

### Description

We already convert measures with agg value MEDIAN to be percentile measures.  Since we're replacing measures with simple metrics, #433 will do the same thing for simple metrics.  Before that, it'd be helpful to have a test in place that shows that the existing legacy measure conversion work, and that's what this PR is for!

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [ ] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)