Clean up Pydantic rules list. (#434)
Towards #387

### Description

We have added several new transformations to support turning measures into metrics, but not all of them made it to our standard rules file.  Our rules file was also very disorganized, and it was hard to tell how things were grouped together or why things were ordered in specific ways.  This is a first pass at all of those issues.

I am hoping to be able to use this in metricflow to organize some of the transformation rules more easily, too, so if I run into problems, there may be further updates to this file.

### Checklist

- [x] I have read [the contributing guide](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md) and understand what's expected of me
- [x] I have signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
- [x] This PR includes tests, or tests are not required/relevant for this PR
- [x] I have run `changie new` to [create a changelog entry](https://github.com/dbt-labs/dbt-semantic-interfaces/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)